dle: Revisit and simplify module structure
