+++ devito/dle/__init__.py
@@ -1,3 +1,5 @@
+from devito.dle.utils import *  # noqa
 from devito.dle.blocking_utils import *  # noqa
+from devito.dle.parallelizer import NThreads, Ompizer  # noqa
+from devito.dle.rewriters import *  # noqa
 from devito.dle.transformer import *  # noqa
-from devito.dle.backends import *  # noqa
