--- devito/dle/backends/__init__.py
@@ -1,6 +1,0 @@
-from devito.dle.backends.common import *  # noqa
-from devito.dle.backends.utils import *  # noqa
-from devito.dle.backends.basic import BasicRewriter  # noqa
-from devito.dle.backends.parallelizer import NThreads, Ompizer  # noqa
-from devito.dle.backends.advanced import (AdvancedRewriter, SpeculativeRewriter,  # noqa
-                                          AdvancedRewriterSafeMath, CustomRewriter)  # noqa
